+++ volmdlr/shells.py
@@ -151,8 +151,15 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
         :return: The corresponding OpenShell3D object.
         :rtype: :class:`volmdlr.faces.OpenShell3D`
         """
+        # Quick fix:
+        # ----------------------------------
+        name = arguments[0][1:-1]
+        if isinstance(arguments[-1], int):
+            product = object_dict[arguments[-1]]
+            name = product[1:-1]
+        # ----------------------------------
         faces = [object_dict[int(face[1:])] for face in arguments[1] if object_dict[int(face[1:])] is not None]
-        return cls(faces, name=arguments[0][1:-1])
+        return cls(faces, name=name)
 
     def to_step(self, current_id):
         """